requirements.txt
asgiref==3.3.1
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.15
django-cors-middleware==1.5.0
django-storages==1.9.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
psycopg2-binary==2.8.6
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
